Address review feedback: simplify disposal and consolidate tests
- DocumentClient.cs: Replace 'is IDisposable' pattern with simple null check
- Move Dispose_DisposesPartitionKeyRangeLocationWhenIDisposable test to
  GlobalPartitionEndpointManagerUnitTests
- Remove Dispose_CancelsCancellationTokenSource, Dispose_IsIdempotent,
  and NoOp_DoesNotImplementIDisposable tests
- Delete DocumentClientDisposeTests.cs (remaining tests undone)
